variables: support duplicate names
DWARF sections technically support having multiple symbols with the same
name. While this is probably an error, it has been observed in the wild and
tasru should handle it.

Add the ability to support arrays of variables with the same name, and add
functions that loop through duplicate names to support providing them to the
user.

Signed-off-by: Sean Cross <sean@xobs.io>